back-end/pyproject.toml
[project]
name = "back-app"
version = "0.1.0"
description = ""
authors = [
    {name = "lucasqtl",email = "lucasquintela04@gmail.com"},
    {name = "Rodrigo Lima",email = "rodrigotenorio2005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.9"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
taskipy = "^1.14.1"
pytest-asyncio = "^0.26.0"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff format . && ruff check --fix .'
run = 'uvicorn back_app.main:app --reload'
pre_test = 'task lint'
test = 'pytest -s --cov=back_app -vv'
post_test = 'coverage html'
